backend/requirements.txt
fastapi[all]
uvicorn
python-multipart
qdrant-client
openai-agents[litellm]
python-dotenv
pytest
fastembed